Moved to files
Moved branch stuff to files along with other things. Flag setters moved to flags.cpp, arm operators moved to arm_operators.h
