+++ SeniorProject2025/SeniorProject2025.vcxproj
@@ -140,11 +140,15 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="branchAndSubroutines.cpp" />
+    <ClCompile Include="flags.cpp" />
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="arm_operators.h" />
+    <ClInclude Include="branchAndSubroutines.h" />
     <ClInclude Include="calculations.h" />
+    <ClInclude Include="flags.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
